style: 优化代码风格
