+++ laokou-common/laokou-common-mybatis-plus/src/main/java/org/laokou/common/mybatisplus/config/TimescaleDB.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright (c) 2022-2026 KCloud-Platform-IoT Author or Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.laokou.common.mybatisplus.config;
+
+import org.laokou.common.storage.AbstractDataSource;
+import org.laokou.common.storage.Config;
+import org.laokou.common.storage.Table;
+import org.springframework.jdbc.core.JdbcTemplate;
+
+import javax.sql.DataSource;
+
+/**
+ * @author laokou
+ */
+public class TimescaleDB extends AbstractDataSource {
+
+	private final TimescaleDBConfig timescaleDBConfig;
+
+	protected JdbcTemplate jdbcTemplate;
+
+	public TimescaleDB(Config config, DataSource dataSource) {
+		this.timescaleDBConfig = (TimescaleDBConfig) config;
+	}
+
+	@Override
+	public void open() {
+
+	}
+
+	@Override
+	public void close() {
+
+	}
+
+	@Override
+	public void create(Table table) {
+
+	}
+
+}
